requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
readability-lxml
python-dotenv
openai